fix(player): fix delegate did not called(issue #8)
WKWebView's property `configuration` is of `@NSCopying` therefore the
original implementation failed to connect `userContentController` with
the targeted scriptMessageHandler. References: https://goo.gl/FM6WbD,
https://goo.gl/AA528r

1. Rewrite the two init(frame:playerVars:) methods so that the

scriptMessageHandler (self) get connected with a local

`userContentController` variable, which is a property of a instance

of `WKWebViewConfiguration` being used to config the behavior of

the initialized YTSwiftyPlayer instance.

2. delete the local variable `userContentController` since it's

no longer used.

3. delete the didSet closure of local variable `delegate` since the

closure is of no use.

Closes #8